feat: lesson 컴포넌트 분리 및 example추가
