+++ src/pages/lesson.tsx
@@ -1,797 +1,39 @@
-import type { NextPage } from "next";
-import Link from "next/link";
-import React, { useRef, useState } from "react";
-import {
-  BigCloseSvg,
-  CloseSvg,
-  DoneSvg,
-  LessonTopBarEmptyHeart,
-  LessonTopBarHeart,
-} from "~/components/Svgs";
-import { useBoundStore } from "~/hooks/useBoundStore";
-import { useRouter } from "next/router";
-
-const lessonProblem1 = {
-  type: "SELECT_1_OF_3",
-  question: `Which one of these is "the apple"?`,
-  answers: [
-    { icon: <LessonTopBarEmptyHeart />, name: "la manzana" },
-    { icon: <LessonTopBarEmptyHeart />, name: "el niño" },
-    { icon: <LessonTopBarEmptyHeart />, name: "la mujer" },
-  ],
-  correctAnswer: 0,
-} as const;
-
-// const lessonProblem2 = {
-//   type: "WRITE_IN_ENGLISH",
-//   question: "El niño",
-//   answerTiles: ["woman", "milk", "water", "I", "The", "boy"],
-//   correctAnswer: [4, 5],
-// } as const;
-
-const lessonProblems = [lessonProblem1, lessonProblem1];
-
-const numbersEqual = (a: readonly number[], b: readonly number[]): boolean => {
-  return a.length === b.length && a.every((_, i) => a[i] === b[i]);
-};
-
-const formatTime = (timeMs: number): string => {
-  const seconds = Math.floor(timeMs / 1000) % 60;
-  const minutes = Math.floor(timeMs / 1000 / 60) % 60;
-  const hours = Math.floor(timeMs / 1000 / 60 / 60);
-  if (hours === 0)
-    return [minutes, seconds]
-      .map((x) => x.toString().padStart(2, "0"))
-      .join(":");
-  return [hours, minutes, seconds]
-    .map((x) => x.toString().padStart(2, "0"))
-    .join(":");
-};
+import { NextPage } from "next";
+import React from "react";
+import TestExample from "~/components/TestExample";
 
 const Lesson: NextPage = () => {
-  const router = useRouter();
+  return (
+    <div>
+      {/* learn페이지에서 버튼 클릭으로 lesson으로 이동함.
+          이 클릭한 문제 번호에 따라 보여주는 컴포넌트를 다르게 가져갈 예정.
+          삼항 삼항 삼항 연산자를 써 
+          유닛 1,2,3,4  
+      */}
 
-  const [lessonProblem, setLessonProblem] = useState(0);
-  const [correctAnswerCount, setCorrectAnswerCount] = useState(0);
-  const [incorrectAnswerCount, setIncorrectAnswerCount] = useState(0);
-  const [selectedAnswer, setSelectedAnswer] = useState<null | number>(null);
-  const [correctAnswerShown, setCorrectAnswerShown] = useState(false);
-  const [quitMessageShown, setQuitMessageShown] = useState(false);
+      <TestExample />
 
-  const [selectedAnswers, setSelectedAnswers] = useState<number[]>([]);
-
-  const startTime = useRef(Date.now());
-  const endTime = useRef(startTime.current + 1000 * 60 * 3 + 1000 * 33);
-
-  const [questionResults, setQuestionResults] = useState<QuestionResult[]>([]);
-  const [reviewLessonShown, setReviewLessonShown] = useState(false);
-
-  const problem = lessonProblems[lessonProblem] ?? lessonProblem1;
-
-  const totalCorrectAnswersNeeded = 2;
-
-  const hearts =
-    "fast-forward" in router.query &&
-    !isNaN(Number(router.query["fast-forward"]))
-      ? 3 - incorrectAnswerCount
-      : null;
-
-  const { correctAnswer } = problem;
-  const isAnswerCorrect = Array.isArray(correctAnswer)
-    ? numbersEqual(selectedAnswers, correctAnswer)
-    : selectedAnswer === correctAnswer;
-
-  const onCheckAnswer = () => {
-    setCorrectAnswerShown(true);
-    if (isAnswerCorrect) {
-      setCorrectAnswerCount((x) => x + 1);
-    } else {
-      setIncorrectAnswerCount((x) => x + 1);
-    }
-    setQuestionResults((questionResults) => [
-      ...questionResults,
-      {
-        question: problem.question,
-        yourResponse: problem.answers[selectedAnswer ?? 0]?.name ?? "",
-        correctResponse: problem.answers[problem.correctAnswer].name,
-      },
-      // 기존 코드 (타입 2개)
-      // {
-      //   question: problem.question,
-      //   yourResponse:
-      //     problem.type === "SELECT_1_OF_3"
-      //       ? problem.answers[selectedAnswer ?? 0]?.name ?? ""
-      //       : selectedAnswers.map((i) => problem.answerTiles[i]).join(" "),
-      //   correctResponse:
-      //     problem.type === "SELECT_1_OF_3"
-      //       ? problem.answers[problem.correctAnswer].name
-      //       : problem.correctAnswer
-      //           .map((i) => problem.answerTiles[i])
-      //           .join(" "),
-      // },
-    ]);
-  };
-
-  const onFinish = () => {
-    setSelectedAnswer(null);
-    setSelectedAnswers([]);
-    setCorrectAnswerShown(false);
-    setLessonProblem((x) => (x + 1) % lessonProblems.length);
-    endTime.current = Date.now();
-  };
-
-  const onSkip = () => {
-    setSelectedAnswer(null);
-    setCorrectAnswerShown(true);
-  };
-
-  if (correctAnswerCount >= totalCorrectAnswersNeeded && !correctAnswerShown) {
-    return (
-      <LessonComplete
-        correctAnswerCount={correctAnswerCount}
-        incorrectAnswerCount={incorrectAnswerCount}
-        startTime={startTime}
-        endTime={endTime}
-        reviewLessonShown={reviewLessonShown}
-        setReviewLessonShown={setReviewLessonShown}
-        questionResults={questionResults}
-      />
-    );
-  }
-
-  switch (problem.type) {
-    case "SELECT_1_OF_3": {
-      return (
-        <ProblemSelect1Of3
-          problem={problem}
-          correctAnswerCount={correctAnswerCount}
-          totalCorrectAnswersNeeded={totalCorrectAnswersNeeded}
-          selectedAnswer={selectedAnswer}
-          setSelectedAnswer={setSelectedAnswer}
-          quitMessageShown={quitMessageShown}
-          correctAnswerShown={correctAnswerShown}
-          setQuitMessageShown={setQuitMessageShown}
-          isAnswerCorrect={isAnswerCorrect}
-          onCheckAnswer={onCheckAnswer}
-          onFinish={onFinish}
-          onSkip={onSkip}
-          hearts={hearts}
-        />
-      );
-    }
-
-    // case "WRITE_IN_ENGLISH": {
-    //   return (
-    //     <ProblemWriteInEnglish
-    //       problem={problem}
-    //       correctAnswerCount={correctAnswerCount}
-    //       totalCorrectAnswersNeeded={totalCorrectAnswersNeeded}
-    //       selectedAnswers={selectedAnswers}
-    //       setSelectedAnswers={setSelectedAnswers}
-    //       quitMessageShown={quitMessageShown}
-    //       correctAnswerShown={correctAnswerShown}
-    //       setQuitMessageShown={setQuitMessageShown}
-    //       isAnswerCorrect={isAnswerCorrect}
-    //       onCheckAnswer={onCheckAnswer}
-    //       onFinish={onFinish}
-    //       onSkip={onSkip}
-    //       hearts={hearts}
-    //     />
-    //   );
-    // }
-  }
+      {/* 
+            구현한 문제 컴포넌트에서 알아서 사용해라. div css구조는 아래와 같이 사용할것.
+            ex)
+            <div className="flex min-h-screen flex-col gap-5 px-4 py-5 sm:px-0 sm:py-0">
+              <div className="flex grow flex-col items-center gap-5">
+                <div className="w-full max-w-5xl sm:mt-8 sm:px-5">
+                  <ProgressBar
+                  correctAnswerCount={correctAnswerCount}
+                  totalCorrectAnswersNeeded={totalCorrectAnswersNeeded}
+                  setQuitMessageShown={setQuitMessageShown}
+                  />
+                </div>
+              </div>
+              <QuitMessage
+                quitMessageShown={quitMessageShown}
+                setQuitMessageShown={setQuitMessageShown}
+              />
+            </div>
+          */}
+    </div>
+  );
 };
 
 export default Lesson;
-
-const ProgressBar = ({
-  correctAnswerCount,
-  totalCorrectAnswersNeeded,
-  setQuitMessageShown,
-  hearts,
-}: {
-  correctAnswerCount: number;
-  totalCorrectAnswersNeeded: number;
-  setQuitMessageShown: (isShown: boolean) => void;
-  hearts: null | number;
-}) => {
-  return (
-    <header className="flex items-center gap-4">
-      {correctAnswerCount === 0 ? (
-        <Link href="/tutorial" className="text-gray-400">
-          <CloseSvg />
-          <span className="sr-only">Exit lesson</span>
-        </Link>
-      ) : (
-        <button
-          className="text-gray-400"
-          onClick={() => setQuitMessageShown(true)}
-        >
-          <CloseSvg />
-          <span className="sr-only">Exit lesson</span>
-        </button>
-      )}
-      <div
-        className="h-4 grow rounded-full bg-gray-200"
-        role="progressbar"
-        aria-valuemin={0}
-        aria-valuemax={1}
-        aria-valuenow={correctAnswerCount / totalCorrectAnswersNeeded}
-      >
-        <div
-          className={
-            "h-full rounded-full bg-blue-500 transition-all duration-700 " +
-            (correctAnswerCount > 0 ? "px-2 pt-1 " : "")
-          }
-          style={{
-            width: `${(correctAnswerCount / totalCorrectAnswersNeeded) * 100}%`,
-          }}
-        >
-          <div className="h-[5px] w-full rounded-full bg-blue-400"></div>
-        </div>
-      </div>
-      {hearts !== null &&
-        [1, 2, 3].map((heart) => {
-          if (heart <= hearts) {
-            return <LessonTopBarHeart key={heart} />;
-          }
-          return <LessonTopBarEmptyHeart key={heart} />;
-        })}
-    </header>
-  );
-};
-
-const QuitMessage = ({
-  quitMessageShown,
-  setQuitMessageShown,
-}: {
-  quitMessageShown: boolean;
-  setQuitMessageShown: (isShown: boolean) => void;
-}) => {
-  return (
-    <>
-      <div
-        className={
-          quitMessageShown
-            ? "fixed bottom-0 left-0 right-0 top-0 z-30 bg-black bg-opacity-60 transition-all duration-300"
-            : "pointer-events-none fixed bottom-0 left-0 right-0 top-0 z-30 bg-black bg-opacity-0 transition-all duration-300"
-        }
-        onClick={() => setQuitMessageShown(false)}
-        aria-label="Close quit message"
-        role="button"
-      ></div>
-
-      <article
-        className={
-          quitMessageShown
-            ? "fixed bottom-0 left-0 right-0 z-40 flex flex-col gap-4 bg-white px-5 py-12 text-center transition-all duration-300 sm:flex-row"
-            : "fixed -bottom-96 left-0 right-0 z-40 flex flex-col bg-white px-5 py-12 text-center transition-all duration-300 sm:flex-row"
-        }
-        aria-hidden={!quitMessageShown}
-      >
-        <div className="flex grow flex-col gap-4">
-          <h2 className="text-lg font-bold sm:text-2xl">
-            Are you sure you want to quit?
-          </h2>
-          <p className="text-gray-500 sm:text-lg">
-            All progress for this lesson will be lost.
-          </p>
-        </div>
-        <div className="flex grow flex-col items-center justify-center gap-4 sm:flex-row-reverse">
-          <Link
-            className="flex w-full items-center justify-center rounded-2xl border-b-4 border-blue-500 bg-blue-400 py-3 font-bold uppercase text-white transition hover:brightness-105 sm:w-48"
-            href="/tutorial"
-          >
-            Quit
-          </Link>
-          <button
-            className="w-full rounded-2xl py-3 font-bold uppercase text-blue-400 transition hover:brightness-90 sm:w-48 sm:border-2 sm:border-b-4 sm:border-gray-300 sm:text-gray-400 sm:hover:bg-gray-100"
-            onClick={() => setQuitMessageShown(false)}
-          >
-            Stay
-          </button>
-        </div>
-      </article>
-    </>
-  );
-};
-
-const CheckAnswer = ({
-  isAnswerSelected,
-  isAnswerCorrect,
-  correctAnswerShown,
-  correctAnswer,
-  onCheckAnswer,
-  onFinish,
-  onSkip,
-}: {
-  isAnswerSelected: boolean;
-  isAnswerCorrect: boolean;
-  correctAnswerShown: boolean;
-  correctAnswer: string;
-  onCheckAnswer: () => void;
-  onFinish: () => void;
-  onSkip: () => void;
-}) => {
-  return (
-    <>
-      <section className="border-gray-200 sm:border-t-2 sm:p-10">
-        <div className="mx-auto flex max-w-5xl sm:justify-between">
-          <button
-            className="hidden rounded-2xl border-2 border-b-4 border-gray-200 bg-white p-3 font-bold uppercase text-gray-400 transition hover:border-gray-300 hover:bg-gray-200 sm:block sm:min-w-[150px] sm:max-w-fit"
-            onClick={onSkip}
-          >
-            Skip
-          </button>
-          {!isAnswerSelected ? (
-            <button
-              className="grow rounded-2xl bg-gray-200 p-3 font-bold uppercase text-gray-400 sm:min-w-[150px] sm:max-w-fit sm:grow-0"
-              disabled
-            >
-              Check
-            </button>
-          ) : (
-            <button
-              onClick={onCheckAnswer}
-              className="grow rounded-2xl border-b-4 border-blue-600 bg-blue-500 p-3 font-bold uppercase text-white sm:min-w-[150px] sm:max-w-fit sm:grow-0"
-            >
-              Check
-            </button>
-          )}
-        </div>
-      </section>
-
-      <div
-        className={
-          correctAnswerShown
-            ? isAnswerCorrect
-              ? "fixed bottom-0 left-0 right-0 bg-blue-100 font-bold text-blue-600 transition-all"
-              : "fixed bottom-0 left-0 right-0 bg-red-100 font-bold text-red-500 transition-all"
-            : "fixed -bottom-52 left-0 right-0"
-        }
-      >
-        <div className="flex max-w-5xl flex-col gap-4 p-5 sm:mx-auto sm:flex-row sm:items-center sm:justify-between sm:p-10 sm:py-14">
-          <>
-            {isAnswerCorrect ? (
-              <div className="mb-2 flex flex-col gap-5 sm:flex-row sm:items-center">
-                <div className="hidden rounded-full bg-white p-5 text-blue-500 sm:block">
-                  <DoneSvg />
-                </div>
-                <div className="text-2xl">Good job!</div>
-              </div>
-            ) : (
-              <div className="mb-2 flex flex-col gap-5 sm:flex-row sm:items-center">
-                <div className="hidden rounded-full bg-white p-5 text-red-500 sm:block">
-                  <BigCloseSvg />
-                </div>
-                <div className="flex flex-col gap-2">
-                  <div className="text-2xl">Correct solution:</div>{" "}
-                  <div className="text-sm font-normal">{correctAnswer}</div>
-                </div>
-              </div>
-            )}
-          </>
-          <button
-            onClick={onFinish}
-            className={
-              isAnswerCorrect
-                ? "w-full rounded-2xl border-b-4 border-blue-600 bg-blue-500 p-3 font-bold uppercase text-white transition hover:brightness-105 sm:min-w-[150px] sm:max-w-fit"
-                : "w-full rounded-2xl border-b-4 border-red-600 bg-red-500 p-3 font-bold uppercase text-white transition hover:brightness-105 sm:min-w-[150px] sm:max-w-fit"
-            }
-          >
-            Continue
-          </button>
-        </div>
-      </div>
-    </>
-  );
-};
-
-const ProblemSelect1Of3 = ({
-  problem,
-  correctAnswerCount,
-  totalCorrectAnswersNeeded,
-  selectedAnswer,
-  setSelectedAnswer,
-  quitMessageShown,
-  correctAnswerShown,
-  setQuitMessageShown,
-  isAnswerCorrect,
-  onCheckAnswer,
-  onFinish,
-  onSkip,
-  hearts,
-}: {
-  problem: typeof lessonProblem1;
-  correctAnswerCount: number;
-  totalCorrectAnswersNeeded: number;
-  selectedAnswer: number | null;
-  setSelectedAnswer: React.Dispatch<React.SetStateAction<number | null>>;
-  correctAnswerShown: boolean;
-  quitMessageShown: boolean;
-  setQuitMessageShown: React.Dispatch<React.SetStateAction<boolean>>;
-  isAnswerCorrect: boolean;
-  onCheckAnswer: () => void;
-  onFinish: () => void;
-  onSkip: () => void;
-  hearts: number | null;
-}) => {
-  const { question, answers, correctAnswer } = problem;
-
-  return (
-    <div className="flex min-h-screen flex-col gap-5 px-4 py-5 sm:px-0 sm:py-0">
-      <div className="flex grow flex-col items-center gap-5">
-        <div className="w-full max-w-5xl sm:mt-8 sm:px-5">
-          <ProgressBar
-            correctAnswerCount={correctAnswerCount}
-            totalCorrectAnswersNeeded={totalCorrectAnswersNeeded}
-            setQuitMessageShown={setQuitMessageShown}
-            hearts={hearts}
-          />
-        </div>
-        <section className="flex max-w-2xl grow flex-col gap-5 self-center sm:items-center sm:justify-center sm:gap-24 sm:px-5">
-          <h1 className="self-start text-2xl font-bold sm:text-3xl">
-            {question}
-          </h1>
-          <div
-            className="grid grid-cols-2 gap-2 sm:grid-cols-3"
-            role="radiogroup"
-          >
-            {answers.map((answer, i) => {
-              return (
-                <div
-                  key={i}
-                  className={
-                    i === selectedAnswer
-                      ? "cursor-pointer rounded-xl border-2 border-b-4 border-blue-300 bg-blue-100 p-4 text-blue-400"
-                      : "cursor-pointer rounded-xl border-2 border-b-4 border-gray-200 p-4 hover:bg-gray-100"
-                  }
-                  role="radio"
-                  aria-checked={i === selectedAnswer}
-                  tabIndex={0}
-                  onClick={() => setSelectedAnswer(i)}
-                >
-                  {answer.icon}
-                  <h2 className="text-center">{answer.name}</h2>
-                </div>
-              );
-            })}
-          </div>
-        </section>
-      </div>
-
-      <CheckAnswer
-        correctAnswer={answers[correctAnswer].name}
-        correctAnswerShown={correctAnswerShown}
-        isAnswerCorrect={isAnswerCorrect}
-        isAnswerSelected={selectedAnswer !== null}
-        onCheckAnswer={onCheckAnswer}
-        onFinish={onFinish}
-        onSkip={onSkip}
-      />
-
-      <QuitMessage
-        quitMessageShown={quitMessageShown}
-        setQuitMessageShown={setQuitMessageShown}
-      />
-    </div>
-  );
-};
-
-// const ProblemWriteInEnglish = ({
-//   problem,
-//   correctAnswerCount,
-//   totalCorrectAnswersNeeded,
-//   selectedAnswers,
-//   setSelectedAnswers,
-//   quitMessageShown,
-//   correctAnswerShown,
-//   setQuitMessageShown,
-//   isAnswerCorrect,
-//   onCheckAnswer,
-//   onFinish,
-//   onSkip,
-//   hearts,
-// }: {
-//   problem: typeof lessonProblem2;
-//   correctAnswerCount: number;
-//   totalCorrectAnswersNeeded: number;
-//   selectedAnswers: number[];
-//   setSelectedAnswers: React.Dispatch<React.SetStateAction<number[]>>;
-//   correctAnswerShown: boolean;
-//   quitMessageShown: boolean;
-//   setQuitMessageShown: React.Dispatch<React.SetStateAction<boolean>>;
-//   isAnswerCorrect: boolean;
-//   onCheckAnswer: () => void;
-//   onFinish: () => void;
-//   onSkip: () => void;
-//   hearts: number | null;
-// }) => {
-//   const { question, correctAnswer, answerTiles } = problem;
-
-//   return (
-//     <div className="flex min-h-screen flex-col gap-5 px-4 py-5 sm:px-0 sm:py-0">
-//       <div className="flex grow flex-col items-center gap-5">
-//         <div className="w-full max-w-5xl sm:mt-8 sm:px-5">
-//           <ProgressBar
-//             correctAnswerCount={correctAnswerCount}
-//             totalCorrectAnswersNeeded={totalCorrectAnswersNeeded}
-//             setQuitMessageShown={setQuitMessageShown}
-//             hearts={hearts}
-//           />
-//         </div>
-//         <section className="flex max-w-2xl grow flex-col gap-5 self-center sm:items-center sm:justify-center sm:gap-24">
-//           <h1 className="mb-2 text-2xl font-bold sm:text-3xl">
-//             Write this in English
-//           </h1>
-
-//           <div className="w-full">
-//             <div className="flex items-center gap-2 px-2">
-//               <Image src={challengeSvg} alt="" width={92} height={115} />
-//               <div className="relative ml-2 w-fit rounded-2xl border-2 border-gray-200 p-4">
-//                 {question}
-//                 <div
-//                   className="absolute h-4 w-4 rotate-45 border-b-2 border-l-2 border-gray-200 bg-white"
-//                   style={{
-//                     top: "calc(50% - 8px)",
-//                     left: "-10px",
-//                   }}
-//                 ></div>
-//               </div>
-//             </div>
-
-//             <div className="flex min-h-[60px] flex-wrap gap-1 border-b-2 border-t-2 border-gray-200 py-1">
-//               {selectedAnswers.map((i) => {
-//                 return (
-//                   <button
-//                     key={i}
-//                     className="rounded-2xl border-2 border-b-4 border-gray-200 p-2 text-gray-700"
-//                     onClick={() => {
-//                       setSelectedAnswers((selectedAnswers) => {
-//                         return selectedAnswers.filter((x) => x !== i);
-//                       });
-//                     }}
-//                   >
-//                     {answerTiles[i]}
-//                   </button>
-//                 );
-//               })}
-//             </div>
-//           </div>
-//           <div className="flex flex-wrap justify-center gap-1">
-//             {answerTiles.map((answerTile, i) => {
-//               return (
-//                 <button
-//                   key={i}
-//                   className={
-//                     selectedAnswers.includes(i)
-//                       ? "rounded-2xl border-2 border-b-4 border-gray-200 bg-gray-200 p-2 text-gray-200"
-//                       : "rounded-2xl border-2 border-b-4 border-gray-200 p-2 text-gray-700"
-//                   }
-//                   disabled={selectedAnswers.includes(i)}
-//                   onClick={() =>
-//                     setSelectedAnswers((selectedAnswers) => {
-//                       if (selectedAnswers.includes(i)) {
-//                         return selectedAnswers;
-//                       }
-//                       return [...selectedAnswers, i];
-//                     })
-//                   }
-//                 >
-//                   {answerTile}
-//                 </button>
-//               );
-//             })}
-//           </div>
-//         </section>
-//       </div>
-
-//       <CheckAnswer
-//         correctAnswer={correctAnswer.map((i) => answerTiles[i]).join(" ")}
-//         correctAnswerShown={correctAnswerShown}
-//         isAnswerCorrect={isAnswerCorrect}
-//         isAnswerSelected={selectedAnswers.length > 0}
-//         onCheckAnswer={onCheckAnswer}
-//         onFinish={onFinish}
-//         onSkip={onSkip}
-//       />
-
-//       <QuitMessage
-//         quitMessageShown={quitMessageShown}
-//         setQuitMessageShown={setQuitMessageShown}
-//       />
-//     </div>
-//   );
-// };
-
-const LessonComplete = ({
-  correctAnswerCount,
-  incorrectAnswerCount,
-  startTime,
-  endTime,
-  reviewLessonShown,
-  setReviewLessonShown,
-  questionResults,
-}: {
-  correctAnswerCount: number;
-  incorrectAnswerCount: number;
-  startTime: React.MutableRefObject<number>;
-  endTime: React.MutableRefObject<number>;
-  reviewLessonShown: boolean;
-  setReviewLessonShown: React.Dispatch<React.SetStateAction<boolean>>;
-  questionResults: QuestionResult[];
-}) => {
-  const router = useRouter();
-  const isPractice = "practice" in router.query;
-
-  const increaseXp = useBoundStore((x) => x.increaseXp);
-  const increaseLessonsCompleted = useBoundStore(
-    (x) => x.increaseLessonsCompleted,
-  );
-  return (
-    <div className="flex min-h-screen flex-col gap-5 px-4 py-5 sm:px-0 sm:py-0">
-      <div className="flex grow flex-col items-center justify-center gap-8 font-bold">
-        <h1 className="text-center text-3xl text-yellow-400">
-          Lesson Complete!
-        </h1>
-        <div className="flex flex-wrap justify-center gap-5">
-          <div className="min-w-[110px] rounded-xl border-2 border-yellow-400 bg-yellow-400">
-            <h2 className="py-1 text-center text-white">Total XP</h2>
-            <div className="flex justify-center rounded-xl bg-white py-4 text-yellow-400">
-              {correctAnswerCount}
-            </div>
-          </div>
-          <div className="min-w-[110px] rounded-xl border-2 border-blue-400 bg-blue-400">
-            <h2 className="py-1 text-center text-white">Committed</h2>
-            <div className="flex justify-center rounded-xl bg-white py-4 text-blue-400">
-              {formatTime(endTime.current - startTime.current)}
-            </div>
-          </div>
-          <div className="min-w-[110px] rounded-xl border-2 border-blue-400 bg-blue-400">
-            <h2 className="py-1 text-center text-white">Amazing</h2>
-            <div className="flex justify-center rounded-xl bg-white py-4 text-blue-400">
-              {Math.round(
-                (correctAnswerCount /
-                  (correctAnswerCount + incorrectAnswerCount)) *
-                  100,
-              )}
-              %
-            </div>
-          </div>
-        </div>
-      </div>
-      <section className="border-gray-200 sm:border-t-2 sm:p-10">
-        <div className="mx-auto flex max-w-5xl sm:justify-between">
-          <button
-            className="hidden rounded-2xl border-2 border-b-4 border-gray-200 bg-white p-3 font-bold uppercase text-gray-400 transition hover:border-gray-300 hover:bg-gray-200 sm:block sm:min-w-[150px] sm:max-w-fit"
-            onClick={() => setReviewLessonShown(true)}
-          >
-            Review lesson
-          </button>
-          <Link
-            className={
-              "flex w-full items-center justify-center rounded-2xl border-b-4 border-blue-600 bg-blue-500 p-3 font-bold uppercase text-white transition hover:brightness-105 sm:min-w-[150px] sm:max-w-fit"
-            }
-            href="/tutorial"
-            onClick={() => {
-              increaseXp(correctAnswerCount);
-              if (!isPractice) {
-                increaseLessonsCompleted();
-              }
-            }}
-          >
-            Continue
-          </Link>
-        </div>
-      </section>
-      <ReviewLesson
-        reviewLessonShown={reviewLessonShown}
-        setReviewLessonShown={setReviewLessonShown}
-        questionResults={questionResults}
-      />
-    </div>
-  );
-};
-
-type QuestionResult = {
-  question: string;
-  yourResponse: string;
-  correctResponse: string;
-};
-
-const ReviewLesson = ({
-  reviewLessonShown,
-  setReviewLessonShown,
-  questionResults,
-}: {
-  reviewLessonShown: boolean;
-  setReviewLessonShown: React.Dispatch<React.SetStateAction<boolean>>;
-  questionResults: QuestionResult[];
-}) => {
-  const [selectedQuestionResult, setSelectedQuestionResult] =
-    useState<null | QuestionResult>(null);
-  return (
-    <div
-      className={[
-        "fixed inset-0 flex items-center justify-center p-5 transition duration-300",
-        reviewLessonShown ? "" : "pointer-events-none opacity-0",
-      ].join(" ")}
-    >
-      <div
-        className={[
-          "absolute inset-0 bg-black",
-          reviewLessonShown ? "opacity-75" : "pointer-events-none opacity-0",
-        ].join(" ")}
-        onClick={() => setReviewLessonShown(false)}
-      ></div>
-      <div className="relative flex w-full max-w-4xl flex-col gap-5 rounded-2xl border-2 border-gray-200 bg-white p-8">
-        <button
-          className="absolute -right-5 -top-5 rounded-full border-2 border-gray-200 bg-gray-100 p-1 text-gray-400 hover:brightness-90"
-          onClick={() => setReviewLessonShown(false)}
-        >
-          <BigCloseSvg className="h-8 w-8" />
-          <span className="sr-only">Close</span>
-        </button>
-        <h2 className="text-center text-3xl">Check out your scorecard!</h2>
-        <p className="text-center text-xl text-gray-400">
-          Click the tiles below to reveal the solutions
-        </p>
-        <div className="grid grid-cols-1 gap-5 sm:grid-cols-2 md:grid-cols-3 lg:grid-cols-4">
-          {questionResults.map((questionResult, i) => {
-            return (
-              <button
-                key={i}
-                className={[
-                  "relative flex flex-col items-stretch gap-3 rounded-xl p-5 text-left",
-                  questionResult.yourResponse === questionResult.correctResponse
-                    ? "bg-yellow-100 text-yellow-600"
-                    : "bg-red-100 text-red-500",
-                ].join(" ")}
-                onClick={() =>
-                  setSelectedQuestionResult((selectedQuestionResult) =>
-                    selectedQuestionResult === questionResult
-                      ? null
-                      : questionResult,
-                  )
-                }
-              >
-                <div className="flex justify-between gap-2">
-                  <h3 className="font-bold">{questionResult.question}</h3>
-                  <div className="flex h-7 w-7 shrink-0 items-center justify-center rounded-full bg-white">
-                    {questionResult.yourResponse ===
-                    questionResult.correctResponse ? (
-                      <DoneSvg className="h-5 w-5" />
-                    ) : (
-                      <BigCloseSvg className="h-5 w-5" />
-                    )}
-                  </div>
-                </div>
-                <div>{questionResult.yourResponse}</div>
-                {selectedQuestionResult === questionResult && (
-                  <div className="absolute left-1 right-1 top-20 z-10 rounded-2xl border-2 border-gray-200 bg-white p-3 text-sm tracking-tighter">
-                    <div
-                      className="absolute -top-2 h-3 w-3 rotate-45 border-l-2 border-t-2 border-gray-200 bg-white"
-                      style={{ left: "calc(50% - 6px)" }}
-                    ></div>
-                    <div className="font-bold uppercase text-gray-400">
-                      Your response:
-                    </div>
-                    <div className="mb-3 text-gray-700">
-                      {questionResult.yourResponse}
-                    </div>
-                    <div className="font-bold uppercase text-gray-400">
-                      Correct response:
-                    </div>
-                    <div className="text-gray-700">
-                      {questionResult.correctResponse}
-                    </div>
-                  </div>
-                )}
-              </button>
-            );
-          })}
-        </div>
-      </div>
-    </div>
-  );
-};
